Added hint for file extension
When loading from data we wouldn't know what file extension the data was coming from so we would always try all loaders. 

Now we try loaders based on the file ext hint to make more educated guesses of the data. If this still fails we fallback to trying all loaders again, but in 99% of the cases the file extension will have the proper data so this speeds up the checking