Moved flowi render interface to separate crate
flowi-core currently drags in quite a bunch of dependencies not needed for the rendering backends to so to separate it into a separate crate means that we can full optimize the rendering backend while keeping most of the other code in debug if we want. It also makes it easier in the future to expand this.