Lots of changes (#12)
* Update to latest Clay layout

* Added mouse, keyboard and gamepad input for SDL2

* WIP on mouse support

* Button hovering implemented

* Started with implementing left menu of launcher

* Arena bug fixes

* Start of left-side menu

* Some fixes related to y text calculations

* Cleanup of vfs, but not fully working yet

* Fixed up text-rendering

* Semi working VFS again

* Fixed another unit-test

* Showing images works

* Removed hard-coded window/tile sizes

* Added a (low quality) image scaler

* Start of demoozoo-fetcher

* cargo fmt

* Added downloading of images

* Changed so no size is needed for loading a font

Instead this is handled internally so that that any size can be
requested from a loaded font which makes it easier to use. We will
evaluate the speed of text generation later so it might make sense to
pre-generated text to speed this up.

* Fixed missing increment of font-id

Also temp-hardcoded so thin gets EXTRA_LIGHT set on it for the fonts

* WIP on background image in main view

* Updated Clay to latest version

* Bug fix for SSE2 memcmp

* Warning fixes

* Fixed incorrect warning change

* Name the background workers

* Reduced dependencies for cosmic-text

* Don't wrap text for now

* Added public function to calc text size.

This needs to be cached later on

* Added initial vignette fade for background images

* WIP on demo launcher UI

* Added access to id_index

* Initial start of demo selection

* WIP on demo launcher prototyping

* More WIP on demo launcher prototype

* Switch to using Color16 for output buffers

It just makes it easier to understand the offsets without needing to use "magic" multipliers for i16 buffer

* Added support for a border around an image

This will allow us to do image sampling "outside" of the visible area
without adding extra checks in the sampling code per pixel. Even if
these checks aren't that expensive, it's annoying to deal with reading
too much with simd instructions as you would have to pick another one.

* Mostly random cleanups while being sick

* Bicubic image sampling ref implementation

* Added support to switch between bilinear and bicubic for image scale

Also did some minor cleanup and fixes

* WIP

* Disable max opt for faster iteration

* Basic fade-out of top row working

* Implemented focusing on items and curve based scrolling

Now the scrolling isn't linear, but curve base instead which gives a much smoother motion. Also to proper support selection if items correct focusing has been implemented now which switch between items.

* Cleanup

* Ignore .idea

* Missed files in selector commit

* Download of party data from demozoo now working

* Show correct competition names instead of dummy

* WIP on remote loading and prio queue

* Simplified cache system

* WIP on cleanup

* WIP on more cleanup

* Code building again after lots of shuffling

* Work on better and simplified IO system + fmt

* WIP

* Loading of demozoo data works again

* Party display back

* Screenshot downloading working

* Deleted unused files

* Some fields can be null from demozoo

* Fixed typo

* More wip on demo selector

* Removed more unused code

* WIP on sharp bilinear

* WIP

* Fixed all warnings

* More cleanup

* Fixed tests

* Use 2021 edition

* ARM / macOS fixes

* Temp removed test

* This breaks windows. Will investigate later